+++ comment-updater-core/src/main/kotlin/org/jetbrains/research/commentupdater/models/config/EmbeddingConfig.kt
