+++ comment-updater-core/src/main/kotlin/org/jetbrains/research/commentupdater/package-info.java
