+++ comment-updater-core/src/main/kotlin/org/jetbrains/research/commentupdater/processors/MethodChangesExtractor.kt
