+++ comment-updater-plugin/build.gradle
@@ -1,0 +1,58 @@
+plugins {
+    id 'java'
+    id 'org.jetbrains.intellij' version '0.7.2'
+    id "org.jetbrains.kotlin.jvm" version "1.4.30"
+}
+
+group = rootProject.group
+version = rootProject.version
+
+    repositories {
+        mavenCentral()
+        jcenter()
+        maven { url "https://www.jetbrains.com/intellij-repository/releases" }
+        maven { url "https://www.jetbrains.com/intellij-repository/snapshots" }
+        maven { url "https://jetbrains.bintray.com/intellij-third-party-dependencies" }
+        maven { url "https://www.myget.org/F/rd-snapshots/maven/" } // Required for VCSTesting
+        maven { url "$projectDir/libs" }
+    }
+
+dependencies {
+
+    implementation(project(":comment-updater-core"))
+
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
+    implementation "io.github.java-diff-utils:java-diff-utils:4.9"
+
+    implementation "io.kinference:inference:0.1.2"
+    implementation "com.microsoft.onnxruntime:onnxruntime:1.8.0"
+    implementation 'com.beust:klaxon:5.5'
+    implementation 'com.github.tsantalis:refactoring-miner:2.1.0'
+    implementation 'org.jetbrains.bio:viktor:1.1.0'
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.0"
+    implementation "org.ojalgo:ojalgo:48.4.0"
+}
+
+// See https://github.com/JetBrains/gradle-intellij-plugin/
+intellij {
+    version = '2021.1'
+    plugins 'git4idea', 'java'
+}
+
+compileKotlin {
+    kotlinOptions.jvmTarget = "1.8"
+}
+compileTestKotlin {
+    kotlinOptions.jvmTarget = "1.8"
+}
+
+
+
+
+
+tasks.runIde {
+    runIde.args "CommentUpdater"
+    runIde.jvmArgs "-Djava.awt.headless=true"
+    maxHeapSize = "20g"
+}
+
