+++ comment-updater-plugin/src/main/kotlin/org/jetbrains/research/commentupdater/package-info.java
@@ -1,0 +1,1 @@
+package org.jetbrains.research.commentupdater;
