+++ comment-updater-plugin/src/main/resources/CommentUpdaterBundle.properties
