+++ comment-updater-plugin/src/main/resources/META-INF/plugin.xml
@@ -1,0 +1,25 @@
+<idea-plugin>
+    <id>org.jetbrains.research.commentupdater</id>
+    <name>CommentUpdater</name>
+    <vendor>url="https://research.jetbrains.org/">JetBrains Research</vendor>
+
+    <description><![CDATA[
+     Detects outdated comments in Java code and suggests updates for them.
+    ]]>
+    </description>
+
+    <depends>com.intellij.modules.java</depends>
+
+    <depends>Git4Idea</depends>
+
+
+    <extensions defaultExtensionNs="com.intellij">
+        <appStarter implementation="PluginRunner"/>
+<!--        <localInspection language="JAVA"-->
+<!--                         displayName="The comment is outdated"-->
+<!--                         groupName="Comment inconsistency"-->
+<!--                         enabledByDefault="true"-->
+<!--                         level="WARNING"-->
+<!--                         implementationClass="CodeCommentInspection"/>-->
+    </extensions>
+</idea-plugin>
