--- libs/com/github/jetbrains/jetCheck/0.2.1/jetCheck-0.2.1.pom
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>com.github.jetbrains</groupId>
-  <artifactId>jetCheck</artifactId>
-  <version>0.2.1</version>
-  <description>POM was created from install:install-file</description>
-</project>
