--- libs/com/github/jetbrains/jetCheck/maven-metadata-local.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<metadata>
-  <groupId>com.github.jetbrains</groupId>
-  <artifactId>jetCheck</artifactId>
-  <versioning>
-    <release>0.2.1</release>
-    <versions>
-      <version>0.2.1</version>
-    </versions>
-    <lastUpdated>20200603200532</lastUpdated>
-  </versioning>
-</metadata>
